test: consolidate embedded tests into api/src/tests/, restrict pytest collection
Move all 42 test functions out of service/route modules into 13 new files
under api/src/tests/, restrict pytest.ini python_files to test_*.py, and
drop the workarounds the old *.py collection forced (--ignore flags,
__test__ = False on the sernia push route handler, which is renamed
send_test_push). database.py's connectivity tests become check_* runtime
health checks with thin test wrappers.

Verified: collect-only parity vs main (488 vs 489 — the one delta is
api/index.py no longer double-collecting test_database_connections via
import), plus a green default suite (two pre-existing local-env failures
behave identically on main).
