--- .github/logfire-investigate/history.log
@@ -1,13 +1,0 @@
-2026-05-31T14:12:52.974289+00:00  fired  session=session_01Jmo3jKJrS6J3L8TAkK9XoP  groups=1  hits=4
-2026-05-31T14:38:51.082256+00:00  fired  session=session_01LFRvFjSXBwgZFL7HGkTKs4  groups=1  hits=4
-2026-06-01T15:06:45.255317+00:00  fired  session=session_01UJi3KgictJ8bjCsfpy1xHF  groups=4  hits=4
-2026-06-05T13:58:51.955448+00:00  fired  session=session_01UWJ39QtJp9qKgvWrmryvKt  groups=1  hits=1
-2026-06-06T13:34:00.794574+00:00  fired  session=session_018BkwGEGL4PLEsfkgie9Lmv  groups=1  hits=1
-2026-06-07T23:36:57.633352+00:00  fired  session=session_01MRCqAxuA83EeTvtUnfT1HL  groups=4  hits=4
-2026-06-08T14:27:14.596947+00:00  fired  session=session_016wZDu6uUXeU4PaYe6CoRCn  groups=4  hits=4
-2026-06-08T23:01:03.758694+00:00  fired  session=session_012z9tZN4KmsaTi5uogZKhap  groups=4  hits=4
-2026-06-21T17:33:04.268488+00:00  fired  session=session_011faXq51tj5WXUZvDEjPnkF  groups=3  hits=3
-2026-06-21T22:52:29.157799+00:00  fired  session=session_01ViuPc4ngWvCBcVvfHQrqmz  groups=3  hits=3
-2026-06-22T14:55:50.970727+00:00  fired  session=session_018KnNB6yWvBScb9VhQ1iLvT  groups=3  hits=3
-2026-06-23T13:53:19.084485+00:00  fired  session=session_01G9wGcudFfDe4etftiFVcYs  groups=1  hits=1
-2026-06-23T22:53:52.564359+00:00  fired  session=session_016fida4zUzU8s8aMinaappR  groups=1  hits=1
