+++ CLAUDE.md
@@ -1,6 +1,6 @@
 # CLAUDE.md - AI Assistant Guide
 
-> **Last Updated**: 2026-06-10
+> **Last Updated**: 2026-07-02
 > `AGENTS.md` is a symlink to this file so non-Claude agent tools pick it up too.
 
 ---
@@ -186,13 +186,11 @@ pnpm approve-builds <package-name>  # Updates pnpm-workspace.yaml allowBuilds
 - **Smoke tests** (`TestSmoke` classes): Fast import/wiring checks that verify modules load correctly and components are connected (e.g., agent has history processors wired, sub-agent models configured). No API keys needed.
 - **Unit tests**: Mock all external calls. Use realistic tool result data matching actual output formats (ClickUp task dumps, Gmail search results, etc.), not dummy strings.
 - **SMS test safety**: NEVER create live tests that send real SMS to external contacts (tenants, vendors, etc.). External SMS tests must ALWAYS mock the send. Only `send_internal_sms` may be tested live against real internal numbers. If a dedicated test phone number is provided in the future, it will be explicitly configured — do not guess or use tenant numbers.
-- **Test location**: `api/src/tests/` (but `pytest.ini` collects `test_*` functions from EVERY `*.py` file under `api/` — see gotchas below).
+- **Test location**: `api/src/tests/`. ALL tests live there — do not embed `test_*` functions inside service modules (that pattern was consolidated away in 2026-07; collection is standard `test_*.py` files only).
 
 ### Pytest Gotchas (learned the hard way)
 
-- `pytest.ini` sets `python_files = *.py`, so inline tests in service modules are collected too. Consequences:
-  - A FastAPI route handler named `test_*` gets collected as a test — set `handler.__test__ = False` (see `sernia_ai/push/routes.py`).
-  - Every package dir under `api/` needs an `__init__.py`, or pytest imports the file under a second top-level module name and SQLAlchemy models get double-registered ("Table already defined").
+- Collection is standard (`python_files = test_*.py *_test.py`, tests in `api/src/tests/`). It used to be `*.py`, which collected tests out of every service module and forced workarounds (`__test__ = False` on FastAPI handlers named `test_*`, `__init__.py` in every package dir to avoid double-import model registration). Don't loosen it back, and avoid naming route handlers `test_*` regardless.
 - **Event loop scope is `session`** (`pytest.ini`): the app uses module-level async engines (asyncpg pools). Per-test loops cause order-dependent "Event loop is closed" failures. Don't change this back to `function`.
 - **Module-level caches leak between tests** — e.g. the tool-result summary cache keyed by `tool_call_id`. Clear them in autouse fixtures when tests reuse short ids (see `test_history_processors.py`).
 - **`TestModel` does not support native tools** (WebSearch/WebFetch). To run `sernia_agent` against `TestModel`, use `with sernia_agent.override(native_tools=[]):` and pass `output_type=str` (the structured output spec needs tool-mode output, which `custom_output_text` can't produce). See `test_sernia_agent_wiring.py`.
