+++ api/src/ai_demos/multi_agent_chat/graph.py
@@ -2,12 +2,10 @@
 Graph definition for routing messages to specialized agents using Pydantic AI Graph Beta API
 """
 
-import json
 import logfire
 from dataclasses import dataclass
 from typing import Literal
 
-import pytest
 from dotenv import load_dotenv
 from pydantic import BaseModel, ConfigDict
 from pydantic_graph.graph_builder import GraphBuilder
@@ -187,80 +185,4 @@ g.add(
 
 # Build the graph
 multi_agent_graph = g.build()
-
-
-@pytest.mark.live
-@pytest.mark.asyncio
-async def test_multi_agent_graph_routes_to_weather():
-    """Test that multi-agent graph routes weather questions to weather agent"""
-    input_data = MultiAgentInput(message="What's the weather in Tokyo?")
-    state = MultiAgentState()
-    result = await multi_agent_graph.run(state=state, inputs=input_data)
-    assert result.agent_name == AgentName.weather
-    assert result.response is not None
-    assert len(result.response) > 0
-    print(result.response)
-
-
-@pytest.mark.live
-@pytest.mark.asyncio
-async def test_multi_agent_graph_routes_to_weather_vercel_ai():
-    """Test that multi-agent graph routes weather questions to weather agent"""
-    input_data = MultiAgentInput(message="What's the weather in Tokyo?")
-    request_payload = {
-        "trigger": "submit-message",
-        "id": "test-id",
-        "messages": [
-            {
-                "id": "msg-1",
-                "role": "user",
-                "parts": [
-                    {
-                        "type": "text",
-                        "text": "What's the weather in Tokyo?"
-                    }
-                ]
-            }
-        ]
-    }
-    vercel_request = build_test_vercel_request(request_payload)
-    state = MultiAgentState(agent_run_method="vercel_ai", vercel_ai_request=vercel_request)
-    result = await multi_agent_graph.run(state=state, inputs=input_data)
-    assert result.agent_name == AgentName.weather
-    assert result.response is not None
-    assert isinstance(result.response, Response)
-    assert result.response.status_code == 200
-    print(result.response)
-
-
-def build_test_vercel_request(payload: dict) -> Request:
-    """Create a Starlette Request suitable for Vercel adapter testing."""
-    body_bytes = json.dumps(payload).encode("utf-8")
-    body_consumed = False
-
-    async def receive():
-        nonlocal body_consumed
-        if body_consumed:
-            return {"type": "http.request", "body": b"", "more_body": False}
-        body_consumed = True
-        return {"type": "http.request", "body": body_bytes, "more_body": False}
-
-    scope = {
-        "type": "http",
-        "asgi": {"version": "3.0", "spec_version": "2.3"},
-        "http_version": "1.1",
-        "method": "POST",
-        "scheme": "http",
-        "path": "/api/ai-demos/multi-agent-chat",
-        "raw_path": b"/api/ai-demos/multi-agent-chat",
-        "root_path": "",
-        "query_string": b"",
-        "headers": [
-            (b"content-type", b"application/json"),
-        ],
-        "client": ("testclient", 50000),
-        "server": ("testserver", 80),
-    }
-
-    return Request(scope, receive)
