+++ api/src/clickup/service.py
@@ -3,7 +3,6 @@ load_dotenv(find_dotenv('.env'))
 import os
 import requests
 from pprint import pprint
-import pytest
 from datetime import datetime
 import pytz
 from api.src.open_phone.service import send_message
@@ -152,8 +151,3 @@ def register_clickup_apscheduler_jobs():
 #
 # def register_clickup_dbos_jobs():
 #     logfire.info("ClickUp DBOS jobs registered.")
-
-@pytest.mark.live
-@pytest.mark.asyncio
-async def test_get_peppino_view_tasks():
-    await get_peppino_view_tasks()
