+++ api/src/google/calendar/service.py
@@ -2,11 +2,8 @@ from __future__ import annotations
 
 import datetime
 from enum import StrEnum
-from pprint import pprint
 
 import logfire
-import pytz
-import pytest
 from fastapi import HTTPException
 from googleapiclient.discovery import build
 from pydantic import BaseModel, EmailStr, Field, field_serializer
@@ -209,55 +206,3 @@ async def delete_calendar_event(service, event_id: str):
         return True
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
-
-
-# ------------------------------------------------------------------------------------------------
-# ------------------------------ TESTS -----------------------------------------------------------
-# ------------------------------------------------------------------------------------------------
-
-
-@pytest.mark.live
-@pytest.mark.asyncio
-async def test_query_calendar_events():
-    service = await get_calendar_service(user_email="emilio@serniacapital.com")
-    now = datetime.datetime.now(tz=pytz.timezone("US/Eastern"))
-    now_str = now.isoformat()
-    print("Getting the upcoming 3 events")
-    events_result = (
-        service.events()
-        .list(
-            calendarId="primary",
-            timeMin=now_str,
-            maxResults=3,
-            singleEvents=True,
-            orderBy="startTime",
-        )
-        .execute()
-    )
-    events = events_result.get("items", [])
-
-    assert len(events) == 3
-
-
-@pytest.mark.live
-@pytest.mark.asyncio
-async def test_create_calendar_event():
-    service = await get_calendar_service(user_email="all@serniacapital.com")
-    et_tz = pytz.timezone("US/Eastern")
-    start_time = datetime.datetime(year=2026, month=3, day=29, hour=12).astimezone(
-        et_tz
-    )
-    end_time = start_time + datetime.timedelta(hours=1)
-
-    event = CalendarEventInput(
-        summary="Test Event3",
-        description="This is a test event",
-        start=start_time,
-        end=end_time,
-        attendees=[CalendarAttendee(email="emilio@serniacapital.com")],
-    )
-    new_event = await create_calendar_event(
-        service, event, organizer_email="all@serniacapital.com", overwrite=True
-    )
-    pprint(new_event)
-    assert new_event["summary"] == event.summary
