+++ api/src/google/common/service_account_auth.py
@@ -2,7 +2,6 @@
 Shared authentication utilities for Google APIs.
 """
 
-import pytest
 from dotenv import load_dotenv, find_dotenv
 load_dotenv(find_dotenv('.env'))
 from google.oauth2 import service_account
@@ -100,11 +99,6 @@ def get_service_credentials(scopes: Optional[List[str]] = None) -> service_accou
             status_code=500,
             detail=f"Failed to initialize Google service account credentials: {str(e)}"
         )
-    
-@pytest.mark.live
-def test_get_service_credentials():
-    creds = get_service_credentials()
-    print(creds)
 
 def get_delegated_credentials(
     user_email: str,
