+++ api/src/google/sheets/service.py
@@ -8,7 +8,6 @@ import os
 from typing import List, Any, Dict
 from fastapi import HTTPException
 import logfire
-import pytest
 
 from api.src.google.common.service_account_auth import get_service_credentials
 
@@ -106,9 +105,4 @@ def get_sheet_as_json(spreadsheet_id: str, sheet_name: str = 'Sheet1') -> List[D
         raise HTTPException(
             status_code=500,
             detail=f"Failed to convert sheet to JSON: {str(e)}"
-        ) 
-
-@pytest.mark.live
-def test_get_contacts_sheet_as_json():
-    spreadsheet_id = '1Gi0Wrkwm-gfCnAxycuTzHMjdebkB5cDt8wwimdYOr_M'
-    return get_sheet_as_json(spreadsheet_id, sheet_name="OpenPhone")
+        )
