+++ pytest.ini
@@ -1,9 +1,11 @@
 [pytest]
-python_files = test_*.py *_test.py *.py
+; Standard collection: tests live in api/src/tests/. (Collection used to be
+; `*.py`, which pulled tests out of every service module and forced hacks
+; like `__test__ = False` on FastAPI handlers named test_* — don't go back.)
+python_files = test_*.py *_test.py
 python_functions = test_*
 ; Filter out live tests by default
-addopts = -s -v -m "not live" --ignore=api/src/database/migrations --ignore=api/src/ai/email_approval_demo/agent.py
-; addopts = -s -v --ignore=api/src/database/migrations --ignore=api/src/ai/email_approval_demo/agent.py
+addopts = -s -v -m "not live"
 markers =
     live: hits real third-party APIs — run explicitly with: pytest -m live
 pythonpath = .
